pyproject.toml
[tool.poetry]
name = "s-kaupat-scraper"
version = "0.1.0"
description = "Scrape s-kaupat.fi → BigQuery"
authors = ["Harri Juntunen <harri@juntunen.ai>"]
readme = "README.md"
packages = [{include = "scraper"}, {include = "loader"}]

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
beautifulsoup4 = "^4.13.4"
playwright = "^1.52.0"
python-dateutil = "^2.9.0.post0"
click = "^8.2.1"
google-cloud-bigquery = "^3.15.0"
google-cloud-monitoring = "^2.15.0"
google-auth = "^2.23.0"
pandas = "^2.1.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pyarrow = "^20.0.0"
undetected-chromedriver = "^3.5.5"
selenium = "^4.33.0"
user-agents = "^2.2.0"
requests = "^2.32.3"
setuptools = "^80.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
